requirements.txt
Flask
Flask-Session
Flask-SQLAlchemy
psycopg2-binary
SQLAlchemy
openai
fpdf
python-dotenv